+++ ui/src/shared/components/GaugeChart.tsx
@@ -1,6 +1,8 @@
 import React, {PureComponent} from 'react'
 import _ from 'lodash'
 
+import {manager} from 'src/worker/JobManager'
+
 import getLastValues from 'src/shared/parsing/lastValues'
 import Gauge from 'src/shared/components/Gauge'
 
@@ -11,9 +13,12 @@ import {ErrorHandling} from 'src/shared/decorators/errors'
 import {DecimalPlaces} from 'src/types/dashboards'
 import {ColorString} from 'src/types/colors'
 import {TimeSeriesServerResponse} from 'src/types/series'
+import {FluxTable} from 'src/types/flux'
+import {DataTypes} from 'src/shared/components/RefreshingGraph'
 
 interface Props {
-  data: TimeSeriesServerResponse[]
+  data: TimeSeriesServerResponse[] | FluxTable[]
+  dataType: DataTypes
   decimalPlaces: DecimalPlaces
   cellID: string
   cellHeight?: number
@@ -23,10 +28,44 @@ interface Props {
   resizerTopHeight?: number
 }
 
+interface State {
+  lastValues?: {
+    values: number[]
+    series: string[]
+  }
+}
+
 @ErrorHandling
-class GaugeChart extends PureComponent<Props> {
+class GaugeChart extends PureComponent<Props, State> {
   public static defaultProps: Partial<Props> = {
     colors: stringifyColorValues(DEFAULT_GAUGE_COLORS),
+  }
+
+  private isComponentMounted: boolean
+
+  constructor(props: Props) {
+    super(props)
+
+    this.state = {}
+  }
+
+  public async componentDidMount() {
+    this.isComponentMounted = true
+    await this.dataToLastValues()
+  }
+
+  public async componentDidUpdate(prevProps: Props) {
+    const isDataChanged =
+      prevProps.dataType !== this.props.dataType ||
+      !_.isEqual(prevProps.data, this.props.data)
+
+    if (isDataChanged) {
+      await this.dataToLastValues()
+    }
+  }
+
+  public componentWillUnmount() {
+    this.isComponentMounted = false
   }
 
   public render() {
@@ -63,15 +102,35 @@ class GaugeChart extends PureComponent<Props> {
   }
 
   private get lastValueForGauge(): number {
-    const {data} = this.props
-    const {lastValues} = getLastValues(data)
-    const lastValue = _.get(lastValues, 0, 0)
+    const {lastValues} = this.state
+    const lastValue = _.get(lastValues, 'values.0', 0)
 
     if (!lastValue) {
       return 0
     }
 
     return lastValue
+  }
+
+  private async dataToLastValues() {
+    const {data, dataType} = this.props
+
+    try {
+      let lastValues
+      if (dataType === DataTypes.flux) {
+        lastValues = await manager.fluxTablesToSingleStat(data as FluxTable[])
+      } else if (dataType === DataTypes.influxQL) {
+        lastValues = getLastValues(data as TimeSeriesServerResponse[])
+      }
+
+      if (!this.isComponentMounted) {
+        return
+      }
+
+      this.setState({lastValues})
+    } catch (err) {
+      console.error(err)
+    }
   }
 }