+++ ui/src/worker/jobs/fluxTablesToDygraph.ts
@@ -3,6 +3,7 @@ import _ from 'lodash'
 import {Message} from 'src/worker/types'
 import {fetchData} from 'src/worker/utils'
 import {FluxTable, DygraphValue} from 'src/types'
+import {parseTablesByTime} from 'src/shared/parsing/flux/parseTablesByTime'
 
 export interface FluxTablesToDygraphResult {
   labels: string[]
@@ -10,75 +11,12 @@ export interface FluxTablesToDygraphResult {
   nonNumericColumns: string[]
 }
 
-const COLUMN_BLACKLIST = new Set([
-  '_time',
-  'result',
-  'table',
-  '_start',
-  '_stop',
-  '',
-])
-
-const NUMERIC_DATATYPES = ['double', 'long', 'int', 'float']
-
 export const fluxTablesToDygraphWork = (
   tables: FluxTable[]
 ): FluxTablesToDygraphResult => {
-  const allColumnNames = []
-  const nonNumericColumns = []
-
-  const tablesByTime = tables.map(table => {
-    const header = table.data[0]
-    const columnNames: {[k: number]: string} = {}
-
-    for (let i = 0; i < header.length; i++) {
-      const columnName = header[i]
-      const dataType = table.dataTypes[columnName]
-
-      if (COLUMN_BLACKLIST.has(columnName)) {
-        continue
-      }
-
-      if (table.groupKey[columnName]) {
-        continue
-      }
-
-      if (!NUMERIC_DATATYPES.includes(dataType)) {
-        nonNumericColumns.push(columnName)
-        continue
-      }
-
-      const uniqueColmnName = Object.entries(table.groupKey).reduce(
-        (acc, [k, v]) => acc + `[${k}=${v}]`,
-        columnName
-      )
-
-      columnNames[i] = uniqueColmnName
-      allColumnNames.push(uniqueColmnName)
-    }
-
-    const timeIndex = header.indexOf('_time')
-
-    if (timeIndex < 0) {
-      throw new Error('Could not find time index in FluxTable')
-    }
-
-    const result = {}
-    for (let i = 1; i < table.data.length; i++) {
-      const row = table.data[i]
-      const time = row[timeIndex]
-
-      result[time] = Object.entries(columnNames).reduce(
-        (acc, [valueIndex, columnName]) => ({
-          ...acc,
-          [columnName]: row[valueIndex],
-        }),
-        {}
-      )
-    }
-
-    return result
-  })
+  const {tablesByTime, allColumnNames, nonNumericColumns} = parseTablesByTime(
+    tables
+  )
 
   const dygraphValuesByTime: {[k: string]: DygraphValue[]} = {}
   const DATE_INDEX = 0