+++ ui/test/shared/components/GaugeChart.test.tsx
@@ -2,6 +2,7 @@ import {shallow} from 'enzyme'
 import React from 'react'
 import Gauge from 'src/shared/components/Gauge'
 import GaugeChart from 'src/shared/components/GaugeChart'
+import {DataTypes} from 'src/shared/components/RefreshingGraph'
 
 const data = [
   {
@@ -30,6 +31,7 @@ const defaultProps = {
     digits: 10,
     isEnforced: false,
   },
+  dataType: DataTypes.influxQL,
 }
 
 const setup = (overrides = {}) => {